Function flac_sys::FLAC__metadata_object_cuesheet_track_resize_indices

pub unsafe extern "C" fn FLAC__metadata_object_cuesheet_track_resize_indices(
    object: *mut FLAC__StreamMetadata, 
    track_num: c_uint, 
    new_num_indices: c_uint
) -> FLAC__bool

Resize a track's index point array.

If the size shrinks, elements will truncated; if it grows, new blank indices will be added to the end.

\param object A pointer to an existing CUESHEET object. \param track_num The index of the track to modify. NOTE: this is not necessarily the same as the track's \a number field. \param new_num_indices The desired length of the array; may be \c 0. \assert \code object != NULL \endcode \code object->type == FLAC__METADATA_TYPE_CUESHEET \endcode \code object->data.cue_sheet.num_tracks > track_num \endcode \code (object->data.cue_sheet.tracks[track_num].indices == NULL && object->data.cue_sheet.tracks[track_num].num_indices == 0) || (object->data.cue_sheet.tracks[track_num].indices != NULL && object->data.cue_sheet.tracks[track_num].num_indices > 0) \endcode \retval FLAC__bool \c false if memory allocation error, else \c true.