Function flac_sys::FLAC__metadata_object_seektable_template_append_points

pub unsafe extern "C" fn FLAC__metadata_object_seektable_template_append_points(
    object: *mut FLAC__StreamMetadata, 
    sample_numbers: *mut FLAC__uint64, 
    num: c_uint
) -> FLAC__bool

Append specific seek point templates to the end of a seek table.

\note As with the other ...seektable_template... functions, you should call FLAC__metadata_object_seektable_template_sort() when finished to make the seek table legal.

\param object A pointer to an existing SEEKTABLE object. \param sample_numbers An array of sample numbers for the seek points. \param num The number of seek point templates to append. \assert \code object != NULL \endcode \code object->type == FLAC__METADATA_TYPE_SEEKTABLE \endcode \retval FLAC__bool \c false if memory allocation fails, else \c true.