Function flac_sys::FLAC__metadata_object_vorbiscomment_entry_to_name_value_pair

pub unsafe extern "C" fn FLAC__metadata_object_vorbiscomment_entry_to_name_value_pair(
    entry: FLAC__StreamMetadata_VorbisComment_Entry, 
    field_name: *mut *mut c_char, 
    field_value: *mut *mut c_char
) -> FLAC__bool

Splits a Vorbis comment entry into NUL-terminated name and value strings.

The returned pointers to name and value will be allocated by malloc() and shall be owned by the caller.

\param entry An existing Vorbis comment entry. \param field_name The address of where the returned pointer to the field name will be stored. \param field_value The address of where the returned pointer to the field value will be stored. \assert \code (entry.entry != NULL && entry.length > 0) \endcode \code memchr(entry.entry, '=', entry.length) != NULL \endcode \code field_name != NULL \endcode \code field_value != NULL \endcode \retval FLAC__bool \c false if memory allocation fails or \a entry does not comply with the Vorbis comment specification, else \c true.