Function flac_sys::FLAC__metadata_object_vorbiscomment_find_entry_from

pub unsafe extern "C" fn FLAC__metadata_object_vorbiscomment_find_entry_from(
    object: *const FLAC__StreamMetadata, 
    offset: c_uint, 
    field_name: *const c_char
) -> c_int

Find a Vorbis comment with the given field name.

The search begins at entry number \a offset; use an offset of 0 to search from the beginning of the comment array.

\param object A pointer to an existing VORBIS_COMMENT object. \param offset The offset into the comment array from where to start the search. \param field_name The field name of the comment to find. \assert \code object != NULL \endcode \code object->type == FLAC__METADATA_TYPE_VORBIS_COMMENT \endcode \code field_name != NULL \endcode \retval int The offset in the comment array of the first comment whose field name matches \a field_name, or \c -1 if no match was found.