Function flac_sys::FLAC__metadata_simple_iterator_init

pub unsafe extern "C" fn FLAC__metadata_simple_iterator_init(
    iterator: *mut FLAC__Metadata_SimpleIterator, 
    filename: *const c_char, 
    read_only: FLAC__bool, 
    preserve_file_stats: FLAC__bool
) -> FLAC__bool

Initialize the iterator to point to the first metadata block in the given FLAC file.

\param iterator A pointer to an existing iterator. \param filename The path to the FLAC file. \param read_only If \c true, the FLAC file will be opened in read-only mode; if \c false, the FLAC file will be opened for edit even if no edits are performed. \param preserve_file_stats If \c true, the owner and modification time will be preserved even if the FLAC file is written to. \assert \code iterator != NULL \endcode \code filename != NULL \endcode \retval FLAC__bool \c false if a memory allocation error occurs, the file can't be opened, or another error occurs, else \c true.