Function flac_sys::FLAC__stream_decoder_get_md5_checking

pub unsafe extern "C" fn FLAC__stream_decoder_get_md5_checking(
    decoder: *const FLAC__StreamDecoder
) -> FLAC__bool

Get the "MD5 signature checking" flag. This is the value of the setting, not whether or not the decoder is currently checking the MD5 (remember, it can be turned off automatically by a seek). When the decoder is reset the flag will be restored to the value returned by this function.

\param decoder A decoder instance to query. \assert \code decoder != NULL \endcode \retval FLAC__bool See above.