Function flac_sys::FLAC__stream_decoder_seek_absolute

pub unsafe extern "C" fn FLAC__stream_decoder_seek_absolute(
    decoder: *mut FLAC__StreamDecoder, 
    sample: FLAC__uint64
) -> FLAC__bool

Flush the input and seek to an absolute sample. Decoding will resume at the given sample. Note that because of this, the next write callback may contain a partial block. The client must support seeking the input or this function will fail and return \c false. Furthermore, if the decoder state is \c FLAC__STREAM_DECODER_SEEK_ERROR, then the decoder must be flushed with FLAC__stream_decoder_flush() or reset with FLAC__stream_decoder_reset() before decoding can continue.

\param decoder A decoder instance. \param sample The target sample number to seek to. \assert \code decoder != NULL \endcode \retval FLAC__bool \c true if successful, else \c false.