Function flac_sys::FLAC__stream_encoder_finish

pub unsafe extern "C" fn FLAC__stream_encoder_finish(
    encoder: *mut FLAC__StreamEncoder
) -> FLAC__bool

Finish the encoding process. Flushes the encoding buffer, releases resources, resets the encoder settings to their defaults, and returns the encoder state to FLAC__STREAM_ENCODER_UNINITIALIZED. Note that this can generate one or more write callbacks before returning, and will generate a metadata callback.

Note that in the course of processing the last frame, errors can occur, so the caller should be sure to check the return value to ensure the file was encoded properly.

In the event of a prematurely-terminated encode, it is not strictly necessary to call this immediately before FLAC__stream_encoder_delete() but it is good practice to match every FLAC__stream_encoder_init_*() with a FLAC__stream_encoder_finish().

\param encoder An uninitialized encoder instance. \assert \code encoder != NULL \endcode \retval FLAC__bool \c false if an error occurred processing the last frame; or if verify mode is set (see FLAC__stream_encoder_set_verify()), there was a verify mismatch; else \c true. If \c false, caller should check the state with FLAC__stream_encoder_get_state() for more information about the error.