Function flac_sys::FLAC__stream_encoder_get_verify_decoder_error_stats

pub unsafe extern "C" fn FLAC__stream_encoder_get_verify_decoder_error_stats(
    encoder: *const FLAC__StreamEncoder, 
    absolute_sample: *mut FLAC__uint64, 
    frame_number: *mut c_uint, 
    channel: *mut c_uint, 
    sample: *mut c_uint, 
    expected: *mut FLAC__int32, 
    got: *mut FLAC__int32
)

Get relevant values about the nature of a verify decoder error. Useful when the stream encoder state is \c FLAC__STREAM_ENCODER_VERIFY_DECODER_ERROR. The arguments should be addresses in which the stats will be returned, or NULL if value is not desired.

\param encoder An encoder instance to query. \param absolute_sample The absolute sample number of the mismatch. \param frame_number The number of the frame in which the mismatch occurred. \param channel The channel in which the mismatch occurred. \param sample The number of the sample (relative to the frame) in which the mismatch occurred. \param expected The expected value for the sample in question. \param got The actual value returned by the decoder. \assert \code encoder != NULL \endcode