Function flac_sys::FLAC__stream_encoder_process

pub unsafe extern "C" fn FLAC__stream_encoder_process(
    encoder: *mut FLAC__StreamEncoder, 
    buffer: *const *const FLAC__int32, 
    samples: c_uint
) -> FLAC__bool

Submit data for encoding. This version allows you to supply the input data via an array of pointers, each pointer pointing to an array of \a samples samples representing one channel. The samples need not be block-aligned, but each channel should have the same number of samples. Each sample should be a signed integer, right-justified to the resolution set by FLAC__stream_encoder_set_bits_per_sample(). For example, if the resolution is 16 bits per sample, the samples should all be in the range [-32768,32767].

For applications where channel order is important, channels must follow the order as described in the frame header.

\param encoder An initialized encoder instance in the OK state. \param buffer An array of pointers to each channel's signal. \param samples The number of samples in one channel. \assert \code encoder != NULL \endcode \code FLAC__stream_encoder_get_state(encoder) == FLAC__STREAM_ENCODER_OK \endcode \retval FLAC__bool \c true if successful, else \c false; in this case, check the encoder state with FLAC__stream_encoder_get_state() to see what went wrong.