Function flac_sys::FLAC__stream_encoder_set_do_exhaustive_model_search

pub unsafe extern "C" fn FLAC__stream_encoder_set_do_exhaustive_model_search(
    encoder: *mut FLAC__StreamEncoder, 
    value: FLAC__bool
) -> FLAC__bool

Set to \c false to let the encoder estimate the best model order based on the residual signal energy, or \c true to force the encoder to evaluate all order models and select the best.

\default \c false \param encoder An encoder instance to set. \param value See above. \assert \code encoder != NULL \endcode \retval FLAC__bool \c false if the encoder is already initialized, else \c true.