Type Definition flac_sys::FLAC__IOCallback_Read

type FLAC__IOCallback_Read = Option<unsafe extern "C" fn(ptr: *mut c_void, size: usize, nmemb: usize, handle: FLAC__IOHandle) -> usize>;

Signature for the read callback. The signature and semantics match POSIX fread() implementations and can generally be used interchangeably.

\param ptr The address of the read buffer. \param size The size of the records to be read. \param nmemb The number of records to be read. \param handle The handle to the data source. \retval size_t The number of records read.