Type Definition flac_sys::FLAC__StreamDecoderErrorCallback

type FLAC__StreamDecoderErrorCallback = Option<unsafe extern "C" fn(decoder: *const FLAC__StreamDecoder, status: FLAC__StreamDecoderErrorStatus, client_data: *mut c_void)>;

Signature for the error callback.

A function pointer matching this signature must be passed to one of the FLAC__stream_decoder_init_*() functions. The supplied function will be called whenever an error occurs during decoding.

\note In general, FLAC__StreamDecoder functions which change the state should not be called on the \a decoder while in the callback.

\param decoder The decoder instance calling the callback. \param status The error encountered by the decoder. \param client_data The callee's client data set through FLAC__stream_decoder_init_*().